Cleave kriging functions from `spatial.py` and move into `kriging.py`
